requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.1
python-magic==0.4.27
pymupdf==1.24.9
python-docx==1.1.2
language-tool-python==2.7.1
spacy==3.7.5
textstat==0.7.4
python-multipart==0.0.9
openai>=1.0.0
python-docx>=1.1.2
openai>=1.40.0,<2
openai==1.77.0
python-dotenv>=1.0
# install the small English model via pip (no runtime downloader):
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
